# src/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorReturn": {
				"properties": {
					"error": {
						"type": "number",
						"format": "double"
					},
					"errorMsg": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductToImageBoddy": {
				"properties": {
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"ImageID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ProductID",
					"ImageID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductToImagesBoddy": {
				"properties": {
					"ProductID": {
						"type": "number",
						"format": "double"
					},
					"ImageIDs": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"ProductID",
					"ImageIDs"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "kevenghomebackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "KevenG"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/edit/product/image": {
			"put": {
				"operationId": "",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"statusbar": {
													"type": "string",
													"enum": [
														"success"
													],
													"nullable": false
												}
											},
											"required": [
												"statusbar"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorReturn"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"edit products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductToImageBoddy"
							}
						}
					}
				}
			}
		},
		"/edit/product/images": {
			"put": {
				"operationId": "",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"statusbar": {
													"type": "string",
													"enum": [
														"success"
													],
													"nullable": false
												}
											},
											"required": [
												"statusbar"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorReturn"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"edit products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "remove",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductToImagesBoddy"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}